+++ src/main/java/com/nc13/StayHo/service/RoomService.java
@@ -9,10 +9,10 @@ import java.util.List;
 public interface RoomService {
     Integer insert(SynthesisDTO params, List<MultipartFile> files);
 
-    List<?> selectByHotel(Long id);
+    List<?> selectByHotel(int id);
     List<?> selectByHotelForSearch(SearchConditionDTO condition);
-    List<?> selectWImg(Long id);
-    SynthesisDTO select(Long id);
-    Integer update(Long id, SynthesisDTO params, List<MultipartFile> files, int[] delImgList);
-    Integer delete(Long[] id);
+    List<?> selectWImg(int id);
+    SynthesisDTO select(int id);
+    Integer update(int id, SynthesisDTO params, List<MultipartFile> files, int[] delImgList);
+    Integer delete(int[] id);
 }
